pyproject.toml
[project]
name = "mcp-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.3",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "wikipedia>=1.4.0",
]
